[INJIMOB-3317]: Fixed compile errors in tests
