+++ kotlin/openID4VP/src/commonMain/kotlin/io/mosip/openID4VP/common/Decoder.kt
@@ -1,3 +1,3 @@
-package io.mosip.openID4VP
+package io.mosip.openID4VP.common
 
 expect fun decodeBase64Data(content: String): ByteArray
