+++ kotlin/openID4VP/src/commonMain/kotlin/io/mosip/openID4VP/common/Encoder.kt
@@ -1,3 +1,3 @@
-package io.mosip.openID4VP
+package io.mosip.openID4VP.common
 
 expect fun encodeToBase64Url(data: ByteArray): String
