+++ kotlin/openID4VP/src/commonTest/kotlin/io/mosip/openID4VP/authorizationRequest/AuthorizationRequestObjectObtainedByReferenceTest.kt
@@ -1,6 +1,5 @@
 package io.mosip.openID4VP.authorizationRequest
 
-import android.util.Log
 import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockkObject
@@ -9,6 +8,7 @@ import io.mockk.verify
 import io.mosip.openID4VP.OpenID4VP
 import io.mosip.openID4VP.authorizationRequest.AuthorizationRequestFieldConstants.*
 import io.mosip.openID4VP.authorizationRequest.AuthorizationRequestFieldConstants.REDIRECT_URI
+import io.mosip.openID4VP.common.Logger
 import io.mosip.openID4VP.common.encodeToJsonString
 import io.mosip.openID4VP.constants.ClientIdScheme
 import io.mosip.openID4VP.constants.ClientIdScheme.*
@@ -48,19 +48,8 @@ class AuthorizationRequestObjectObtainedByReferenceTest {
             )
         } returns mapOf("body" to didResponse)
 
-        mockkStatic(android.util.Log::class)
-        every { Log.e(any(), any()) } answers {
-            val tag = arg<String>(0)
-            val msg = arg<String>(1)
-            println("Error: logTag: $tag | Message: $msg")
-            0
-        }
-        every { Log.d(any(), any()) } answers {
-            val tag = arg<String>(0)
-            val msg = arg<String>(1)
-            println("Error: logTag: $tag | Message: $msg")
-            0
-        }
+        mockkObject(Logger)
+        every { Logger.error(any(), any(), any()) } answers {  }
     }
 
     @After
