+++ kotlin/openID4VP/src/commonTest/kotlin/io/mosip/openID4VP/authorizationRequest/presentationDefinition/FieldsTest.kt
@@ -1,11 +1,13 @@
 package io.mosip.openID4VP.authorizationRequest.presentationDefinition
 
-import android.util.Log
+
 import io.mockk.clearAllMocks
 import io.mockk.every
+import io.mockk.mockkObject
 import io.mockk.mockkStatic
 import io.mosip.openID4VP.authorizationRequest.deserializeAndValidate
 import io.mosip.openID4VP.authorizationRequest.exception.AuthorizationRequestExceptions.InvalidInputPattern
+import io.mosip.openID4VP.common.Logger
 import io.mosip.openID4VP.exceptions.Exceptions
 import io.mosip.openID4VP.exceptions.Exceptions.MissingInput
 
@@ -20,13 +22,8 @@ class FieldsTest {
 
 	@Before
 	fun setUp() {
-		mockkStatic(Log::class)
-		every { Log.e(any(), any()) } answers {
-			val tag = arg<String>(0)
-			val msg = arg<String>(1)
-			println("Error: logTag: $tag | Message: $msg")
-			0
-		}
+        mockkObject(Logger)
+        every { Logger.error(any(), any(), any()) } answers {  }
 	}
 
 	@After
