+++ kotlin/openID4VP/src/jvmMain/kotlin/io/mosip/openID4VP/common/Decoder.jvm.kt
@@ -1,4 +1,4 @@
-package io.mosip.openID4VP
+package io.mosip.openID4VP.common
 
 actual fun decodeBase64Data(content: String): ByteArray {
     return java.util.Base64.getUrlDecoder().decode(content.toByteArray())
